Add files via upload
This is an example blender project that you can use as a starting point. Alternatively, you can put this file in your asset folder so that you will have access to all the PSX style materials, camera settings, and the model size limit object. Although, if you are starting in a blank file you'll need to copy the render settings I have set up for "Color Management" to get the PSX look in the viewport.